+++ packages/@runlightyear/airtable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/airtable
 
+## 1.2.19
+
+### Patch Changes
+
+- Updated dependencies [b0b12552d]
+  - @runlightyear/lightyear@2.2.0
+
 ## 1.2.18
 
 ### Patch Changes