+++ packages/@runlightyear/gmail/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gmail
 
+## 0.6.20
+
+### Patch Changes
+
+- Updated dependencies [b0b12552d]
+  - @runlightyear/lightyear@2.2.0
+
 ## 0.6.19
 
 ### Patch Changes