+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/postmark
 
+## 0.5.19
+
+### Patch Changes
+
+- Updated dependencies [b0b12552d]
+  - @runlightyear/lightyear@2.2.0
+
 ## 0.5.18
 
 ### Patch Changes